[WIP] - changed time type from long to double (further changes is may be required)

+++ src/main/java/jabs/scenario/NormalEthereumNetworkScenario.java
@@ -9,7 +9,7 @@ import static jabs.event.EventFactory.createBlockGenerationEvents;
 import static jabs.event.EventFactory.createTxGenerationEvents;
 
 public class NormalEthereumNetworkScenario extends AbstractScenario {
-    long simulationTime = 0;
+    double simulationTime = 0;
 
     private final int numOfMiners;
     private final int numOfNonMiners;
